requirements.txt
selenium
pytest
allure-pytest
PyGithub
pytest-ordering
pydantic
facebook_business
requests
pandas